Make Sigma a builtin operator
Move the Sigma operator into the builtins and drop the standalone plugin target. The operator now compiles Sigma rules to TQL2 expressions and uses the neo filtering path instead of legacy pattern operands.

Remove the obsolete Sigma unit tests and rely on the existing neo integration coverage for the operator.
