+++ test/tests/operators/sigma/condition.txt
@@ -1,0 +1,15 @@
+{
+  event: {
+    id: "alpha",
+  },
+}
+{
+  event: {
+    id: "bravo",
+  },
+}
+{
+  event: {
+    id: "ignored",
+  },
+}
