+++ test/tests/operators/sigma/directory.tql
@@ -1,3 +1,3 @@
 from {foo: "needle"}
-sigma f"{env("TENZIR_INPUTS")}/sigma/directory"
+sigma f"{env("TENZIR_INPUTS")}/directory"
 sort rule.title
