+++ test/tests/operators/sigma/inputs/basic-rule.yaml
