+++ test/tests/operators/sigma/inputs/condition.yaml
@@ -1,0 +1,11 @@
+title: condition sigma rule
+detection:
+  selection_alpha:
+    foo: alpha
+  selection_bravo:
+    bar: bravo
+  selection_charlie:
+    baz: charlie
+  filter:
+    skip: true
+  condition: 1 of selection_* and not filter
