+++ test/tests/operators/sigma/inputs/directory/alpha.yml
