+++ test/tests/operators/sigma/inputs/directory/nested/bravo.yaml
