+++ test/tests/operators/sigma/inputs/modifiers.yaml
@@ -1,0 +1,12 @@
+title: modifier sigma rule
+detection:
+  numeric:
+    count|gte: 42
+    count|lt: 50
+  cidr:
+    ip|cidr: 192.168.0.0/24
+  encoded:
+    payload|base64: value
+  offset:
+    url|base64offset|contains: http://
+  condition: numeric or cidr or encoded or offset
