+++ test/tests/operators/sigma/refresh_interval_zero.tql
@@ -3,4 +3,4 @@ error: true
 ---
 
 from {foo: "needle"}
-sigma f"{env("TENZIR_INPUTS")}/sigma/basic-rule.yaml", refresh_interval=0s
+sigma f"{env("TENZIR_INPUTS")}/basic-rule.yaml", refresh_interval=0s
